refactor : 응답 구조 추가 개선
