+++ src/main/java/site/codemonster/comon/domain/article/dto/response/ArticleResponse.java
@@ -1,6 +1,5 @@
 package site.codemonster.comon.domain.article.dto.response;
 
-import site.codemonster.comon.domain.article.entity.Article;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 import java.time.LocalDateTime;
@@ -14,14 +13,4 @@ public record ArticleResponse(
     String memberName,
     String memberImage
 ) {
-    public static ArticleResponse of(Article article) {
-        return new ArticleResponse(
-            article.getArticleId(),
-            article.getArticleTitle(),
-            article.getArticleBody(),
-            article.getCreatedDate(),
-            article.getMember().getMemberName(),
-            article.getMember().getImageUrl()
-        );
-    }
 }
